+++ src/Client/ClientCore/Rendering/RenderingManager.cs
@@ -79,7 +79,7 @@ public class RenderingManager : IStartable
     {
         SelectConfiguration();
         CreateMainDisplay();
-        InitializeGUI();
+        InitializeGui();
         LoadResources();
         InitializeRenderer();
     }
@@ -111,6 +111,7 @@ public class RenderingManager : IStartable
     /// </summary>
     public void Render()
     {
+        guiManager.NewFrame();
         renderer.Render();
     }
 
@@ -172,7 +173,7 @@ public class RenderingManager : IStartable
     /// <summary>
     ///     Initializes the GUI.
     /// </summary>
-    private void InitializeGUI()
+    private void InitializeGui()
     {
         try
         {