+++ src/Client/VeldridRenderer/Rendering/VeldridRenderer.cs
@@ -20,6 +20,7 @@ using Castle.Core.Logging;
 using Sovereign.ClientCore.Rendering;
 using Sovereign.ClientCore.Rendering.Configuration;
 using Sovereign.ClientCore.Rendering.Scenes;
+using Sovereign.VeldridRenderer.Rendering.Gui;
 using Sovereign.VeldridRenderer.Rendering.Resources;
 using Veldrid;
 
@@ -34,6 +35,8 @@ public class VeldridRenderer : IRenderer
     ///     Veldrid graphics device.
     /// </summary>
     private readonly VeldridDevice device;
+
+    private readonly GuiRenderer guiRenderer;
 
     /// <summary>
     ///     Veldrid resource manager.
@@ -56,12 +59,13 @@ public class VeldridRenderer : IRenderer
     private bool isDisposed;
 
     public VeldridRenderer(VeldridDevice device, VeldridResourceManager resourceManager,
-        SceneManager sceneManager, VeldridSceneConsumer sceneConsumer)
+        SceneManager sceneManager, VeldridSceneConsumer sceneConsumer, GuiRenderer guiRenderer)
     {
         this.device = device;
         this.resourceManager = resourceManager;
         this.sceneManager = sceneManager;
         this.sceneConsumer = sceneConsumer;
+        this.guiRenderer = guiRenderer;
     }
 
     public ILogger Logger { private get; set; } = NullLogger.Instance;
@@ -78,6 +82,9 @@ public class VeldridRenderer : IRenderer
 
             /* Initialize all scenes. */
             sceneConsumer.Initialize();
+
+            // Initialize additional rendering layers.
+            guiRenderer.Initialize();
         }
         catch (Exception e)
         {
@@ -89,6 +96,7 @@ public class VeldridRenderer : IRenderer
     {
         if (!isDisposed)
         {
+            guiRenderer.Dispose();
             sceneConsumer.Dispose();
             resourceManager.Dispose();
             device.Dispose();
@@ -103,10 +111,10 @@ public class VeldridRenderer : IRenderer
         if (resourceManager.CommandList == null)
             throw new InvalidOperationException("Command list not ready.");
 
+        var commandList = resourceManager.CommandList;
         try
         {
             /* Prepare for rendering. */
-            var commandList = resourceManager.CommandList;
             commandList.Begin();
             commandList.SetFramebuffer(device.Device.SwapchainFramebuffer);
             commandList.ClearColorTarget(0, RgbaFloat.Black);
@@ -117,15 +125,19 @@ public class VeldridRenderer : IRenderer
             sceneConsumer.ConsumeScene(scene);
             scene.EndScene();
 
-            /* Render and present the next frame. */
-            commandList.End();
-            device.Device.SubmitCommands(commandList);
-            device.Device.WaitForIdle();
-            device.Device.SwapBuffers();
+            // GUI rendering from Dear ImGui.
+            guiRenderer.RenderGui(commandList);
         }
         catch (Exception e)
         {
             Logger.Error("Error during rendering.", e);
         }
+
+        /* Render and present the next frame. */
+        commandList.End();
+        device.Device.SubmitCommands(commandList);
+        device.Device.WaitForIdle();
+        device.Device.SwapBuffers();
+        guiRenderer.EndFrame();
     }
 }